Revive integration tests for nasl syntax.
Jira: SC-1183
