+++ rust/tests/data/crash-prefix-recursion.nasl
