+++ rust/tests/data/crash-recursion-depth.nasl
